Codec.MIME.QuotedPrintable

Description

To and from QP content encoding.

Documentation

decode :: String -> String

decode incoming quoted-printable content, stripping out soft line breaks and translating =XY sequences into their decoded byte/octet. The output encoding/representation is still a String, not a sequence of bytes.

encode :: String -> String

encode converts a sequence of characeter _octets_ into quoted-printable form; suitable for transmission in MIME payloads. Note the stress on _octets_; it is assumed that you have already converted Unicode into a <=8-bit encoding (UTF-8, most likely.)